+++ src/main/java/me/kall/camelextend/api/CamelEdible.java
@@ -1,6 +1,0 @@
-package me.kall.camelextend.api;
-
-public interface CamelEdible {
-    boolean camelExtend$isCamelEdible();
-    void camelExtend$setIsCamelEdible(boolean isCamelEdible);
-}