+++ src/main/java/me/kall/camelextend/mixin/CamelAiMixin.java
@@ -3,11 +3,16 @@ package me.kall.camelextend.mixin;
 import com.google.common.collect.ImmutableList;
 import com.mojang.datafixers.util.Pair;
 import me.kall.camelextend.behavior.EatEdibleBlockBehavior;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.world.entity.ai.memory.MemoryModuleType;
+import net.minecraft.world.entity.animal.camel.Camel;
 import net.minecraft.world.entity.animal.camel.CamelAi;
 import org.jetbrains.annotations.NotNull;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.Redirect;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(CamelAi.class)
 public abstract class CamelAiMixin {
@@ -15,5 +20,13 @@ public abstract class CamelAiMixin {
     @Redirect(method = "initIdleActivity", at = @At(value = "INVOKE", target = "Lcom/google/common/collect/ImmutableList;of(Ljava/lang/Object;Ljava/lang/Object;Ljava/lang/Object;Ljava/lang/Object;Ljava/lang/Object;Ljava/lang/Object;)Lcom/google/common/collect/ImmutableList;", remap = false))
     private static <E> @NotNull ImmutableList<E> init(E e1, E e2, E e3, E e4, E e5, E e6) {
         return ImmutableList.of((E) Pair.of(0, new EatEdibleBlockBehavior()), e1, e2, e3, e4, e5, e6);
+    }
+
+    @Mixin(CamelAi.RandomSitting.class)
+    public abstract static class RandomSittingMixin {
+        @Inject(method = "checkExtraStartConditions(Lnet/minecraft/server/level/ServerLevel;Lnet/minecraft/world/entity/animal/camel/Camel;)Z", at = @At("HEAD"), cancellable = true)
+        private void startable(ServerLevel level, @NotNull Camel owner, CallbackInfoReturnable<Boolean> cir) {
+            if (owner.getBrain().hasMemoryValue(MemoryModuleType.WALK_TARGET)) cir.setReturnValue(Boolean.FALSE);
+        }
     }
 }