+++ src/main/java/me/kall/camelextend/mixin/dmg/CamelMixin.java
@@ -1,4 +1,4 @@
-package me.kall.camelextend.mixin;
+package me.kall.camelextend.mixin.dmg;
 
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.damagesource.DamageTypes;
@@ -8,7 +8,7 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(Camel.class)
-public class CamelMixin extends LivingEntityMixin {
+public abstract class CamelMixin extends LivingEntityMixin {
     @Override
     protected void onHurt(@NotNull DamageSource source, float amount, CallbackInfoReturnable<Boolean> cir) {
         if (source.is(DamageTypes.CACTUS)) cir.setReturnValue(false);