+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-# 1.0.3
-Make cactus search an idle activity
-Camel will not search cactus when panic
+# 1.1.0
+- Rewrite using my Duplicationless lib for cleaner codes and better perf
+- Original config is removed but now you can configure camel edible blocks via datapack! See data/camelextend/tags/blocks/camel_edible.json